requirements.txt
torch
torchvision
tensorboard
matplotlib
numpy
pillow
requests
